requirements.txt
streamlit
rdkit-pypi
scikit-learn
matplotlib
pandas
numpy
umap-learn
xgboost==1.5.0
py3Dmol
stmol